add srsi

+++ src/TradingApp.TradingAdapter/Constants/RsiSettingsConst.cs
@@ -2,7 +2,7 @@
 
 public static class RsiSettingsConst
 {
-    public const decimal Overbought = 60;
-    public const decimal Oversold = -60;
+    public const double Overbought = 60;
+    public const double Oversold = -60;
     public const int DefaultPeriod = 14;
 }
